remove sqlite3 log, one middleqare export
