+++ index.js
@@ -1,7 +1,7 @@
 const app = require("./server");
-const logger = require("./logger");
+const { logger } = require("./middleware");
 
 const port = process.env.PORT || 3001;
 app.listen(port, () => {
   logger.info(`Server listening on port ${port}`)
-})
+});
