--- logger/index.js
@@ -1,13 +1,0 @@
-const pino = require('pino');
-
-const logger = pino({
-    transport: {
-        target: 'pino-pretty',
-        options: {
-            translateTime: 'SYS:dd-mm-yyyy HH:MM:ss',
-            ignore: 'pid, hostname'
-        }
-    }
-})
-
-module.exports = logger;
