+++ middleware/index.js
@@ -1,3 +1,14 @@
+const pino = require('pino');
+
+const logger = pino({
+    transport: {
+        target: 'pino-pretty',
+        options: {
+            translateTime: 'SYS:dd-mm-yyyy HH:MM:ss',
+            ignore: 'pid, hostname'
+        }
+    }
+})
 
 const checkRequestBody = (req, res, next) => {
     const { text } = req.body;
@@ -11,4 +22,4 @@ const checkRequestBody = (req, res, next) => {
     next();
 }
 
-module.exports = {  checkRequestBody };
+module.exports = { logger, checkRequestBody };
